pyproject.toml
[project]
name = "test-fast-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Devis",email = "davi.ribeirodosanjos@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.116.2,<0.117.0)",
    "pydantic[email] (>=2.11.9,<3.0.0)",
    "sqlmodel (>=0.0.25,<0.0.26)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "redis (>=6.4.0,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
